checks for associated courses in cce_weightages and observation_groups in cce_grade_set and prevents if any from deleting
git-svn-id: file:///svn/fedena2.1/branches/fedena2.1_i18n@3354 1b8ff446-76f8-11e0-a33c-ba5893e5f458

+++ app/controllers/cce_weightages_controller.rb
@@ -45,10 +45,15 @@ class CceWeightagesController < ApplicationController
 
   def destroy
     @weightage=CceWeightage.find(params[:id])
-    if @weightage.destroy
-      flash[:notice]="Weightage deleted."
+
+    if @weightage.courses.empty?
+      if @weightage.destroy
+        flash[:notice]="Weightage deleted."
+      else
+        flash[:warn_notice]="Weightage could be deleted."
+      end
     else
-      flash[:notice]="Weightage could be deleted."
+      flash[:warn_notice]="CCE weightage #{@weightage.weightage}(#{@weightage.criteria_type}) has been assigned to courses. Remove the associations before deleting."
     end
     redirect_to :action => "index"
   end
